requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
openai==1.35.5

# TESTS
pytest==8.2.2